backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery[redis]==5.3.6
redis==5.0.1
websockets==12.0
langchain==0.1.0
langchain-community==0.0.9
langchain-core==0.1.10
langgraph==0.0.20
httpx==0.25.2
python-multipart==0.0.6
faker==20.1.0
google-generativeai>=0.8.0
